# backend/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Complete API documentation for the ticket management system",
    "title": "Ticket Management System API",
    "contact": {},
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
...
    "/ticket": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Create a new ticket (accessible to all authenticated users)",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Tickets"],
        "summary": "Create new ticket",
        "parameters": [
          {
            "description": "Ticket creation details",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "description": {
                  "type": "string"
                },
                "skills": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "title": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "schema": {
              "$ref": "#/definitions/handlers.TicketSummaryResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/ticket/all": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Retrieve a list of all tickets",
        "produces": ["application/json"],
        "tags": ["Tickets"],
        "summary": "Get all tickets",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/handlers.TicketSummaryResponse"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/ticket/assigned": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Retrieve tickets assigned to the current user",
        "produces": ["application/json"],
        "tags": ["Tickets"],
        "summary": "Get assigned tickets",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/handlers.TicketSummaryResponse"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/ticket/number/{number}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Retrieve ticket by its ticket number",
        "produces": ["application/json"],
        "tags": ["Tickets"],
        "summary": "Get ticket by number",
        "parameters": [
          {
            "type": "integer",
            "description": "Ticket Number",
            "name": "number",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handlers.TicketResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/ticket/stats": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get ticket statistics based on user role (admin/agent see all, users see own)",
        "produces": ["application/json"],
        "tags": ["Tickets"],
        "summary": "Get ticket statistics",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handlers.TicketStats"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/ticket/{id}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Retrieve full details of a specific ticket",
        "produces": ["application/json"],
        "tags": ["Tickets"],
        "summary": "Get ticket by ID",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "Ticket UUID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handlers.TicketResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Permanently delete a ticket (admin/agent only for tickets not created by them)",
        "produces": ["application/json"],
        "tags": ["Tickets"],
        "summary": "Delete ticket",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "Ticket UUID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Partially update ticket fields. Valid state transitions: open→pending/cancelled, pending→open/resolved/cancelled, resolved→open/pending/closed/cancelled",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Tickets"],
        "summary": "Update ticket",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "Ticket UUID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Fields to update",
            "name": "request",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "assigned_to": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "description": {
                  "type": "string"
                },
                "priority": {
                  "type": "string"
                },
                "skills": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "state": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/handlers.TicketSummaryResponse"
            }
          },
          "400": {
            "description": "Invalid state transition or invalid priority",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "handlers.TicketResponse": {
      "type": "object",
      "properties": {
        "assigned_to": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": ["550e8400-e29b-41d4-a716-446655440000"]
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "example": "2025-01-22T10:30:00Z"
        },
        "created_by": {
          "type": "string",
          "format": "uuid",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "creator": {
          "$ref": "#/definitions/handlers.UserInfo"
        },
        "description": {
          "type": "string",
          "example": "Users cannot login with SSO"
        },
        "id": {
          "type": "string",
          "format": "uuid",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "priority": {
          "type": "string",
          "example": "high"
        },
        "skills": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": ["golang", "postgresql"]
        },
        "state": {
          "type": "string",
          "example": "open"
        },
        "ticket_number": {
          "type": "integer",
          "example": 1001
        },
        "title": {
          "type": "string",
          "example": "Fix login bug"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "example": "2025-01-22T11:45:00Z"
        }
      }
    },
    "handlers.TicketStats": {
      "type": "object",
      "properties": {
        "mine": {
          "type": "integer"
        },
        "open": {
          "type": "integer"
        },
        "pending": {
          "type": "integer"
        },
        "resolved": {
          "type": "integer"
        },
        "total": {
          "type": "integer"
        }
      }
    },
    "handlers.TicketSummaryResponse": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string",
          "format": "date-time",
          "example": "2025-01-22T10:30:00Z"
        },
        "description": {
          "type": "string",
          "example": "Users cannot login with SSO"
        },
        "id": {
          "type": "string",
          "format": "uuid",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "priority": {
          "type": "string",
          "example": "high"
        },
        "state": {
          "type": "string",
          "example": "open"
        },
        "ticket_number": {
          "type": "integer",
          "example": 1001
        },
        "title": {
          "type": "string",
          "example": "Fix login bug"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time",
          "example": "2025-01-22T11:45:00Z"
        }
      }
    },
    "handlers.UserInfo": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "format": "email",
          "example": "john.doe@example.com"
        },
        "first_name": {
          "type": "string",
          "example": "John"
        },
        "id": {
          "type": "string",
          "format": "uuid",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "last_name": {
          "type": "string",
          "example": "Doe"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Type \"Bearer\" followed by a space and JWT token.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}